requirements.txt
Flask==3.0.0
openai-whisper==20231117
sentence-transformers==2.2.2
scikit-learn==1.3.2
nltk==3.8.1
numpy==1.24.3
requests==2.31.0
torch==2.1.0
joblib==1.3.2
Werkzeug==3.0.1